requirements.txt
yt-dlp
youtube-transcript-api
tqdm
textblob
pyyaml
